+++ idea/testData/spring/core/gutter/innerBean/InnerBean.kt
@@ -1,0 +1,1 @@
+class Inner<caret>Bean