+++ idea/testData/spring/core/gutter/methodTypeMultiple/Bean.kt
@@ -1,0 +1,4 @@
+class Bean {
+    fun my<caret>Fun() {
+    }
+}